Implemented lock deletion
I wrote the tests for deleting a lock that doesn't exist and a lock that
does. I also removed some code that repeated from creaing locks in the
form of verifying if a LockName was valid and enursing a lock exists or
does not exist. I also added a comment pointing out that it would be
nice to give my exceptions more descriptive names. I'm thinking
something along the lines of "validationException". The more I see the
NoSuchAlgorithmsExceptions the more I want it to be unchecked.