Changed method names and implemented acquireSequencer
I looked over some of the naming conventions for locks in the JDK and
found that the acquire/release convention makes more semantic sense than
GetLock and GetSequencer. I was also able to implement the new
AcquireSequencer method fairly easily.

When thinking of naming I believe I will probably change the name of
KeepLockAlive to just "KeepAlive" since it's just for sequencers in
general and not just locks. Next time will probably be a little test
engineering and renaming and then it'll be back to implementation.