Start some test engineering
Moved on to enhancing some of my tests. Cutting down on some boilerplate
with some methods that can create the locks and trying out a custom
matcher for some of my repeat sets of error requirements. I had some
difficulty adapting this because I use the assertThatThrownBy in my
tests instead of extracting the exception object. I'm going to check the
assertj source next time to see how and if they use
AbstractThrowableAssert to create a specific-Exception
assertThatThrownBy or if I'll have to change the tests a bit.