Use custom matcher for exceptions
A shorter session today, but a helpful one. I was able to complete the
matcher for the StatusRuntimeExceptions that are thrown by the server. I
made a generic matcher for the lock name being invalid and it appears to
have worked. I think further testing might be in order (or at least some
debugging) but I'm happy with it how it is and I think it will make the
tests more readable, shorter, and more consistent. Next time I'll be
doing more test engineering. Hopefully I'll be finishing this part off.