Finish test engineering
I got everything I wanted for now done in the tests. I can see some
possible future changes, but for now this is a good foundation for how I
want the tests to be. The should be readable and the test methods I'm
using will be a great place to start for the java client. I'll be moving
on to writing tests for the next method which should be acquire lock.