Wrote comments for acquireLock and test ideas
I forgot how much the acquire lock logic does for the server. It's a
full keep-alive plus pass fail on the passed in sequencer. Definitely a
good example of the separation between the server logic and just the
lock logic. Next time I'll be writing out the tests for the acquireLock
method.