Add validation checks for submitted sequencers
This time I wanted to work on the lock/unlock, but I realized that I
needed some kind of basic validation of the sequencer before doing
anything. Verification is heavy and checking anything before that is
cheap because it's just a read. It'll be onward and upward for the lock
method next time.