Implemented KeepAlive with tests
I also added a new assertion object for sequencers. It's much cleaner
than it was before. Release lock should be the next and final method to
implement for the core functionality (non-introspective functionality).
I think I might take sec to clean up the server code, because the error
handling is getting a little messy and there's a lot of things done
multiple times that can be fixed up. Logging is also an important part
of the server that I haven't event touched yet. I think it might be
better to tackle those things after the base functionality is done and
before the introspection stuff is implemented.