Revise locks to be less messy with exceptions
I made it so the digest instance isn't created everytime a hash key has
to be made. I added it to the class so the user of the class can manage
the exception and what to do about it. This allows some intesting
options with possible fallback algorithms if your first choice of digest
algorithm isn't available. The keep alive interval will be next and then
we'll see what we can do for the failing tests in server.