Remove references to Guava
The project has a dependency on grpc, but I would like some of the
server implementation classes to stay agnostic. Either way, it wasn't
good to bring in a second version of guava that could potentially
interfere with the version of guava that grpc is bringing in. I want to
clean up some more todos, but the next big thing would be for me to test
the thread safety of the locks. Initially this would be a basic test,
but later I'll make some test scenarios around it.