Complete server cleanup
Some methods were removed, but mostly things were broken out into helper
accessor methods or helper exception methods. Ensuring a lock exists
before doing anything to it is valuable even if we have to protect
methods later on from concurrency side-effects. The methods as they are
now are much cleaner and take concurrently side affects into account.
Moving forward there is still combing through the Lock class to see if
it's also thread safe. Once that is done we can move on to deletion
tokens and the introspection methods