Fix issues with double locking and double unlocking
The two atomic operations is not ideal, but there's only one atomic
mutation and the previous state is always going to happen before the
atomic update. I couldn't think of a way (besides using side effects) to
make sure that there's just one operation, so that will have to do for
now. I'm going to move on to making things more thread safe.