Make sure the sequence number is added as soon as possible
This is a half measure and kind of defeats the purpose of having
explicitly numbered sequencers. The solution to this is to use a true
head instead of a queue that could potentially incorrectly order
sequencers. Even a priority queue could have issues if the
createSequencer finished for sequencer 3 before sequencer 2.