Remove blocking queue for sequencer tracking
The only way we'll keep the sequencer ordering secure is to have a head
pointer that increments to the next head. And to have that head update
using an atomic reference. Interested to test this with threads and see
what breaks. Debating whether to tackle introspection yet or if I should
try out the lock creation returning a deletion key thing. I think that
would be a fairly major re-work, so I'll do the introspection instead
and then tackle the reworks.