Add a destroy token to create/destroy locks
This way only the user who created the lock or someone they give the
token to, can destroy a lock. It would be pretty silly if someone made a
rogue worker that would just list all locks and destroy all locks so
this is a nice touch. I think this is my last real added feature to the
server. Next up is some test engineering and exception tracking.