Made assertion objects for locks
Still working on some test engineering for now. I'll nest the classes in
internal objects to group them by the methods they're testing and then
I'll do some debugging to see what is causing those big exceptions in my
tests. The exceptions don't seem to be a problem for my tests, but they
are noisy so it would be nice to see what's causing them.