Test for null sequencers
This should wrap up things that I want to do server side for now in
terms of functionality. The next step is to extract the implementation
into an interface along with interfaces for the locks and lock names.
I'll also make sure the CLI works to start the server properly as well.
That might require more research on my end about grpc channel types